+++ src/main/java/carpet_client/CarpetClient.java
@@ -2,9 +2,8 @@ package carpet_client;
 
 import carpet.CarpetExtension;
 import carpet.CarpetServer;
-import carpet.CarpetSettings;
 import carpet_client.network.ServerMessageHandler;
-import carpet_client.utils.CarpetRules;
+import carpet_client.utils.CarpetSettingsClientNetworkHandler;
 import carpet_client.utils.Reference;
 import com.mojang.brigadier.CommandDispatcher;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
@@ -49,7 +48,7 @@ public class CarpetClient implements CarpetExtension
         // reloading of /carpet settings is handled by carpet
         // reloading of own settings is handled as an extension, since we claim own settings manager
         Reference.isCarpetClientPresent = true;
-        CarpetRules.attachServer(server);
+        CarpetSettingsClientNetworkHandler.attachServer(server);
     }
 
     @Override
