+++ src/main/java/carpet_client/gui/ServerRulesScreen.java
@@ -1,6 +1,7 @@
 package carpet_client.gui;
 
-import carpet_client.utils.CarpetRules;
+import carpet_client.utils.CarpetSettingsClientNetworkHandler;
+import carpet_client.utils.CarpetSettingsServerNetworkHandler;
 import com.mojang.blaze3d.platform.GlStateManager;
 import net.minecraft.client.gui.screen.Screen;
 import net.minecraft.client.gui.widget.ButtonWidget;
@@ -31,7 +32,7 @@ public class ServerRulesScreen extends Screen
         this.list = new ConfigListWidget(this, this.minecraft);
         this.children.add(this.list);
         this.addButton(new ButtonWidget(this.width / 2 - 155, this.height - 29, 150, 20, "Refresh", (buttonWidget) -> {
-            CarpetRules.requestUpdate(this.minecraft);
+            CarpetSettingsServerNetworkHandler.requestUpdate(this.minecraft);
         }));
         this.addButton(new ButtonWidget(this.width / 2 - 155 + 160, this.height - 29, 150, 20, I18n.translate("gui.done"), (buttonWidget) -> {
             this.minecraft.openScreen(this.parent);
