+++ src/main/java/carpet_client/utils/CarpetSettingsClientNetworkHandler.java
@@ -2,29 +2,26 @@ package carpet_client.utils;
 
 import carpet.CarpetServer;
 import carpet_client.gui.ConfigScreen;
-import carpet_client.network.ClientMessageHandler;
-import io.netty.buffer.Unpooled;
-import net.minecraft.client.MinecraftClient;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.nbt.ListTag;
 import net.minecraft.nbt.Tag;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.util.PacketByteBuf;
 
-public class CarpetRules
+public class CarpetSettingsClientNetworkHandler
 {
     public static PacketByteBuf data;
     private static MinecraftServer server;
     
     public static void setAllData(PacketByteBuf buf)
     {
-        CarpetRules.data = buf;
+        CarpetSettingsClientNetworkHandler.data = buf;
         split();
     }
     
     public static void attachServer(MinecraftServer server)
     {
-        CarpetRules.server = server;
+        CarpetSettingsClientNetworkHandler.server = server;
     }
     
     private static void split()
@@ -45,26 +42,10 @@ public class CarpetRules
         Reference.isCarpetServer = true;
     }
     
-    public static void ruleData(PacketByteBuf data)
+    public static void updateRule(PacketByteBuf data)
     {
         String rule = data.readString();
         String newValue = data.readString();
         CarpetServer.settingsManager.getRule(rule).set(server.getCommandSource(), newValue);
-    }
-    
-    public static void ruleChange(String rule, String newValue, MinecraftClient client)
-    {
-        PacketByteBuf data = new PacketByteBuf(Unpooled.buffer());
-        data.writeVarInt(Reference.RULE_REQUEST);
-        data.writeString(rule);
-        data.writeString(newValue);
-        ClientMessageHandler.sendPacket(data, client);
-    }
-    
-    public static void requestUpdate(MinecraftClient client)
-    {
-        PacketByteBuf data = new PacketByteBuf(Unpooled.buffer());
-        data.writeVarInt(Reference.ALL_GUI_INFO);
-        ClientMessageHandler.sendPacket(data, client);
     }
 }
